refactor: Refactoring of the library to make it compatible with app builder and to modify the default handler.
The library was modified to allow the use of dependencies both when run with a code builder, like Vitejs, for example, or without it. The default handler has also been modified so that if element is a link, the click event is fired instead of using a redirect. This allows links that do not point to a url, but use javascript, to be used.